LayoutTests/http/tests/media/hls/track-in-band-multiple-cues-expected.txt


Test that multiple cues in a sample buffer are parsed correctly.

EVENT(addtrack)
EVENT(cuechange)
EXPECTED (track.cues.length === '5') OK

END OF TEST
